refactor(workspace): centralize builtin front plugins
